move text utils to separate file
